bloglist loaded from api
