+++ src/model/index.js
@@ -1,0 +1,3 @@
+import BlogPost from './BlogPost';
+
+export { BlogPost };
