+++ src/pages/home/Blog.jsx
@@ -1,5 +1,5 @@
 import { useParams, useNavigate } from "react-router-dom";
-import Card from "./Card";
+import Card from "../../components/Card";
 
 export default function Blog() {
   const { id } = useParams();
